Add NDG linker script changes
The NDG application framework make use of custom sections in the final
binary add the required sections to the ARC and X86 linker scripts.

This is an interim change until we define a mechanism to allow
application developers to add sections to the binary without requiring
modification of the base linker scripts.

*** INTERNAL ONLY ***

Signed-off-by: Dirk Brandewie <dirk.j.brandewie@intel.com>
Change-Id: I4282537fbf31bc121a1ddc258da57fa5bfea907f

Conflicts:
	include/nanokernel/arc/v2/scripts/atlas_peak/linker.cmd