atlas_peak-arc: fix interrupt storm issue with printk
JIRA: FIRE-752

The current way of implementing printk from ARC with the use of
Mailboxes lead to interrupt storms on the LMT, and overflow of the
K_swapper event stack. (LMT stays in Mailbox interrupt, services timer
interrupts and add a TICK event int the K_swapper event stack) but the
K_swapper fiber is never run as we are servicing ARC printk interrupts.

This patch fixes this problem by avoiding interrupt storms. for each new
line characters sent to lakemont, the ARC will run a busy loop in order
to give a chance to LMT to go out of interrupt context..

Change-Id: Id65fee6542225600de3073be7f661a054bef3550
Signed-off-by: Alexandre d'Alton <alexandre.d'alton@intel.com>
Reviewed-on: https://android.intel.com/371231
Reviewed-by: Cheverneuil, FredericX <fredericx.cheverneuil@intel.com>
Tested-by: Cheverneuil, FredericX <fredericx.cheverneuil@intel.com>