rtc: Introduce QMI RTC device driver
This patch introduces the 'QMI RTC device driver' which is simply a shim
driver based on RTC driver provided by QMI BSP. This is an experimental
driver and the Designware RTC driver have precedence. So, to enable this
driver, the Designware RTC driver shouldn't be selected.

Finally, since we now have two RTC drivers, this patch also adds the
"Real-Time Clock" menu.

Change-Id: I9f73c58017ab2a1eef042b4e5c08205dbe27b501
Signed-off-by: Andre Guedes <andre.guedes@intel.com>