pinmux: wrap _set/_get in debug ifdefs
The pinmux_set and pinmux_get functions should NOT be standard
functionality for a pinmux.  Their use creates an unstable base
platform for debug and reproducability.  The functions are
useful for cases on early on exploration and testing, but any
final product should set the values in the *_pinmux.c file.

Change-Id: I71850a9526fac40a05a1bca558e2351bf8ef88e2
Signed-off-by: Dan Kalowsky <daniel.kalowsky@intel.com>