gpio: Add QMI-based implementation
Introduce the GPIO QMI-based implementation. This is basically a
shim layer that implement's Zephyr's GPIO APIs on top of QMI drivers.

This is an experimental driver that conflicts with the previous
GPIO_DW implementation. In order to enable it you must set:
- CONFIG_GPIO_DW=n
- CONFIG_QMI_GPIO=y
- CONFIG_GPIO_DW_0=y
- CONFIG_QMI_DRIVERS=y
- CONFIG_LIBQMI_PATH="PATH_TO_QMI"

Note that this driver currently only supports one controller
instance, GPIO_0, so you must select CONFIG_GPIO_DW_0=y. It is
implemented this way due to a limitation from the current version
of QMI.

Missing:
- support both callback use-cases (per pin and per port);
- support multiple controller instances (gpio_0, gpio_1, etc);
- implement gpio_resume() and gpio_suspend() callbacks;
- enable level triggered interrupts in sync with system clock,
through setting INT_CLOCK_SYNC properly.

Change-Id: Ie04d14e2eb8a9a38d706687d53a24d92291100b5
Signed-off-by: Jesus Sanchez-Palencia <jesus.sanchez-palencia@intel.com>