atlaspeak: fix broken ARC boot
We now have an atp_ss_0 device which starts up the ARC instead of
calling it from atp_init().

The shared memory block is now in a header and has a field for status
flags set by the ARC, so that we can ensure that not only is the CPU
running, but Zephyr has initialized on it.

Some changes were made to tiny-mountain-app's flash layout without
updating how Zephyr was linked, with the fixes herein ARC will now
boot.

RAM was improperly set up as ARC and LMT share an address space.
Reserve 1k at the beginning for shared data such as the reset
vector, 24K for ARC, 55K for LMT.

Change-Id: Ib576359f000da305b4a1ce2b2f93815059055176
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>