QuarkSE init: allows step by step gdb execution on arc.
Allows to configure that the QuarkSE processor
does not start the arc processor after the initialization using
the ARC_INIT_DEBUG config variable.

This allows the developer to start kernel execution with GDB
and run the arc core step by step from the beginning.

Otherwise, the arc core is started by QuarkSE and GDB
cannot engage on a step by step execution.

Change-Id: I5d5bcf82b4351c283992637e4b6087e479c2e9fe
Signed-off-by: Juan Manuel Cruz <juan.m.cruz.alcaraz@linux.intel.com>