rtc: Introduce QMSI RTC device driver
This patch introduces the 'QMSI RTC device driver' which is simply a shim
driver based on RTC driver provided by QMSI BSP. This is an experimental
driver and the Designware RTC driver have precedence. So, to enable this
driver, the Designware RTC driver shouldn't be selected.

Finally, since we now have two RTC drivers, this patch also adds the
"Real-Time Clock" menu.

Change-Id: I99d8f6b3cf12d28d2910f3ca471bccbd559702d1
Signed-off-by: Andre Guedes <andre.guedes@intel.com>
Signed-off-by: Anas Nashif <anas.nashif@intel.com>