gpio: Add QMSI-based implementation
Introduce the GPIO QMSI-based implementation. This is basically a
shim layer that implement's Zephyr's GPIO APIs on top of QMSI drivers.

This is an experimental driver that conflicts with the previous
GPIO_DW implementation. In order to enable it you must set:
- CONFIG_GPIO_DW=n
- CONFIG_QMSI_GPIO=y
- CONFIG_GPIO_DW_0=y
- CONFIG_QMSI_DRIVERS=y
- CONFIG_LIBQMSI_PATH="PATH_TO_QMSI"

Note that this driver currently only supports one controller
instance, GPIO_0, so you must select CONFIG_GPIO_DW_0=y. It is
implemented this way due to a limitation from the current version
of QMSI.

Missing:
- support both callback use-cases (per pin and per port);
- support multiple controller instances (gpio_0, gpio_1, etc);
- implement gpio_resume() and gpio_suspend() callbacks;
- enable level triggered interrupts in sync with system clock,
through setting INT_CLOCK_SYNC properly.

Change-Id: Ib61b153dae9741806a9a31d7dc1f82b96d000fbe
Signed-off-by: Jesus Sanchez-Palencia <jesus.sanchez-palencia@intel.com>
Signed-off-by: Anas Nashif <anas.nashif@intel.com>