qmsi: Fix build when using qm_flash.c
On 1.1 Beta RC3, qm_flash.c depends on qm_interrupt.c due to
calls for qm_irq_enable() and qm_irq_disable(). These calls should be
harmless, and they are recommend by the flash controller's datasheet,
but we will need to investigate further if this needs a fix on QMSI
itself.

For now, if the QMSI Flash driver is used, then we must build
qm_interrupt.c .

Change-Id: I98c3036a37108881ca382c46780982a819ddd69c
Signed-off-by: Jesus Sanchez-Palencia <jesus.sanchez-palencia@intel.com>