tracing: add trace points to kernel
Add hooks in key locations in the kernel to enable tracing of key
events. The tracing hooks can be implemented to support various backends
such as SEGGER systemview, Percepio Tracealyzer and othe tools.

In some places we replace the kernel event logger hooks with the new
tracing hooks. The kernel event logger will be removed in favor of this.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>