kernel: mutex: add error checking
- k_mutex_unlock will now perform error checking and return on failures.
- k_mutex_init now will not unconditionally reset the mutex if it is
owned by someone else.
- Add k_mutex_reset for resetting a mutex if it s not busy.

k_mutex_unlock:

If the current thread does not own the mutex, we will now return -EPERM.
In the unlikely situation where we own a lock and the lock count is
zero, we assert. This is considered an undefined bahviour and should not
happen.

k_mutex_init:
Do not reset a mutex object if it is owned by some other thread.

k_mutex_reset:
Reset a mutex if it is not busy.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>