add new CTF bottom: buffered and extensible, optionally formatted
Zephyr decides to use CTF to support kinds of tracing tool, so put the
generic infrastructure for collection of tracing data under ctf bottom.
Now it will buffer tracing packet, add the tracing packet to tracing
list and finally tracing core outputs the tracing packet with the help
of tracing backend and in this commit uart backend is added.
And unbuffered version will be developed based on this infrastructure.

Signed-off-by: Wentong Wu <wentong.wu@intel.com>