benchmarks: don't typecast timing value in arch common code
The variables to store timer values in the benchmark timing code
are of uint64_t but are being type casted to uint32_t before
storing. This type casting is not even consistent as value is
type casted to uint64_t before storing in
arch_timing_value_swap_common. Just remove all the type cast.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>