drivers: dw_intc: cleanup ace intc drivers
ace headers were defining the same structure for the designware
interrupt driver as the existing driver, use the structure from
intc_dw.h and drop the duplicate definitions from the ace register
header.

Additionally, we will be getting the register from DTS directly and not
from the header, so store register in the driver config and use it from
there. This brings the driver closer to the existing DW driver we use
for CAVS.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>